+++ bundle/src/InternalConfig.php
@@ -2,10 +2,22 @@
 
 namespace Hyvor\Internal\Bundle;
 
+use Hyvor\Internal\Component\Component;
+
 class InternalConfig
 {
 
     public function __construct(
+
+        /**
+         * This is APP_KEY in laravel and APP_SECRET in symfony
+         * It is in the Laravel format: base64:<key>
+         */
+        private readonly string $appSecret,
+
+        /**
+         * Component name
+         */
         private readonly string $component,
         private readonly string $instance,
         private readonly ?string $privateInstance,
@@ -13,9 +25,20 @@ class InternalConfig
     ) {
     }
 
-    public function getComponent(): string
+    // returns the app secret with the base64: prefix
+    public function getAppSecretRaw(): string
     {
-        return $this->component;
+        return $this->appSecret;
+    }
+
+    public function getAppSecret(): string
+    {
+        return base64_decode(substr($this->appSecret, 7));
+    }
+
+    public function getComponent(): Component
+    {
+        return Component::from($this->component);
     }
 
     public function getInstance(): string
@@ -26,6 +49,11 @@ class InternalConfig
     public function getPrivateInstance(): ?string
     {
         return $this->privateInstance;
+    }
+
+    public function getPrivateInstanceWithFallback(): string
+    {
+        return $this->privateInstance ?? $this->instance;
     }
 
     public function isFake(): bool