+++ src/Auth/Auth.php
@@ -2,9 +2,9 @@
 
 namespace Hyvor\Internal\Auth;
 
-use Hyvor\Internal\InternalApi\ComponentType;
+use Hyvor\Internal\Component\Component;
+use Hyvor\Internal\Component\ComponentUrlResolver;
 use Hyvor\Internal\InternalApi\Exceptions\InternalApiCallFailedException;
-use Hyvor\Internal\InternalApi\InstanceUrl;
 use Hyvor\Internal\InternalApi\InternalApi;
 use Hyvor\Internal\InternalApi\InternalApiMethod;
 use Illuminate\Http\RedirectResponse;
@@ -37,7 +37,7 @@ class Auth
         }
 
         $response = $this->internalApi->call(
-            ComponentType::CORE,
+            Component::CORE,
             InternalApiMethod::POST,
             '/auth/check',
             [
@@ -73,7 +73,7 @@ class Auth
             urlencode($redirectUrl);
 
         return redirect(
-            InstanceUrl::getInstanceUrl() .
+            ComponentUrlResolver::getInstanceUrl() .
             '/' .
             $page .
             $redirect
@@ -104,7 +104,7 @@ class Auth
     protected function getUsersByField(string $field, iterable $values): Collection
     {
         $response = InternalApi::call(
-            ComponentType::CORE,
+            Component::CORE,
             InternalApiMethod::POST,
             '/auth/users/from/' . $field,
             [