+++ src/Billing/BillingFake.php
@@ -3,13 +3,13 @@
 namespace Hyvor\Internal\Billing;
 
 use Hyvor\Internal\Billing\License\License;
-use Hyvor\Internal\InternalApi\ComponentType;
+use Hyvor\Internal\Component\Component;
 
 class BillingFake extends Billing
 {
 
     /**
-     * @param License|(callable(int $userId, ?int $blogId, ComponentType $component) : ?License)|null $license
+     * @param License|(callable(int $userId, ?int $blogId, Component $component) : ?License)|null $license
      * @return void
      */
     public static function enable(
@@ -22,15 +22,15 @@ class BillingFake extends Billing
 
     public function __construct(
         /**
-         * @param License|(callable(int $userId, ?int $resouceId, ComponentType $component) : ?License)|null $license
+         * @param License|(callable(int $userId, ?int $resouceId, Component $component) : ?License)|null $license
          */
         private readonly mixed $license = null
     ) {
     }
 
-    public function license(int $userId, ?int $resourceId, ?ComponentType $component = null): ?License
+    public function license(int $userId, ?int $resourceId, ?Component $component = null): ?License
     {
-        $component ??= ComponentType::current();
+        $component ??= Component::current();
 
         if ($this->license === null) {
             return null;