+++ tests/Unit/InternalApi/ComponentTypeTest.php
@@ -5,72 +5,72 @@ namespace Hyvor\Internal\Tests\Unit\InternalApi;
 use Hyvor\Internal\Billing\License\Plan\BlogsPlan;
 use Hyvor\Internal\Billing\License\Plan\CorePlan;
 use Hyvor\Internal\Billing\License\Plan\TalkPlan;
-use Hyvor\Internal\InternalApi\ComponentType;
+use Hyvor\Internal\Component\Component;
 use Hyvor\Internal\Tests\TestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
 
-#[CoversClass(ComponentType::class)]
+#[CoversClass(Component::class)]
 class ComponentTypeTest extends TestCase
 {
 
     public function testName(): void
     {
-        $this->assertEquals('HYVOR', ComponentType::CORE->name());
-        $this->assertEquals('Hyvor Talk', ComponentType::TALK->name());
-        $this->assertEquals('Hyvor Blogs', ComponentType::BLOGS->name());
+        $this->assertEquals('HYVOR', Component::CORE->name());
+        $this->assertEquals('Hyvor Talk', Component::TALK->name());
+        $this->assertEquals('Hyvor Blogs', Component::BLOGS->name());
     }
 
     public function testFromConfig(): void
     {
         config(['internal.component' => 'core']);
-        $this->assertEquals(ComponentType::current(), ComponentType::CORE);
+        $this->assertEquals(Component::current(), Component::CORE);
 
         config(['internal.component' => 'talk']);
-        $this->assertEquals(ComponentType::current(), ComponentType::TALK);
+        $this->assertEquals(Component::current(), Component::TALK);
 
         config(['internal.component' => 'blogs']);
-        $this->assertEquals(ComponentType::current(), ComponentType::BLOGS);
+        $this->assertEquals(Component::current(), Component::BLOGS);
     }
 
     public function testGetCoreUrl(): void
     {
         config(['internal.instance' => 'https://hyvor.com']);
-        $this->assertEquals(ComponentType::CORE->getCoreUrl(), 'https://hyvor.com');
+        $this->assertEquals(Component::CORE->getCoreUrl(), 'https://hyvor.com');
 
         config(['internal.instance' => 'https://talk.hyvor.com']);
-        $this->assertEquals(ComponentType::TALK->getCoreUrl(), 'https://hyvor.com');
+        $this->assertEquals(Component::TALK->getCoreUrl(), 'https://hyvor.com');
 
         // externl
         config(['internal.instance' => 'https://hyvor.mycompany.com']);
-        $this->assertEquals(ComponentType::CORE->getCoreUrl(), 'https://hyvor.mycompany.com');
+        $this->assertEquals(Component::CORE->getCoreUrl(), 'https://hyvor.mycompany.com');
 
         // external product
         config(['internal.instance' => 'https://talk.hyvor.mycompany.com']);
-        $this->assertEquals(ComponentType::TALK->getCoreUrl(), 'https://hyvor.mycompany.com');
+        $this->assertEquals(Component::TALK->getCoreUrl(), 'https://hyvor.mycompany.com');
     }
 
     public function testGetTheUrl(): void
     {
         // core
-        $this->assertEquals(ComponentType::CORE->getUrlOf(ComponentType::TALK), 'https://talk.hyvor.com');
-        $this->assertEquals(ComponentType::CORE->getUrlOf(ComponentType::CORE), 'https://hyvor.com');
+        $this->assertEquals(Component::CORE->getUrlOf(Component::TALK), 'https://talk.hyvor.com');
+        $this->assertEquals(Component::CORE->getUrlOf(Component::CORE), 'https://hyvor.com');
 
         // product
-        $this->assertEquals(ComponentType::TALK->getUrlOf(ComponentType::CORE), 'https://hyvor.com');
-        $this->assertEquals(ComponentType::TALK->getUrlOf(ComponentType::BLOGS), 'https://blogs.hyvor.com');
+        $this->assertEquals(Component::TALK->getUrlOf(Component::CORE), 'https://hyvor.com');
+        $this->assertEquals(Component::TALK->getUrlOf(Component::BLOGS), 'https://blogs.hyvor.com');
 
         // other subdomain
         config(['internal.instance' => 'https://hyvor.mycompany.com']);
-        $this->assertEquals(ComponentType::BLOGS->getUrlOf(ComponentType::CORE), 'https://hyvor.mycompany.com');
-        $this->assertEquals(ComponentType::BLOGS->getUrlOf(ComponentType::TALK), 'https://talk.hyvor.mycompany.com');
+        $this->assertEquals(Component::BLOGS->getUrlOf(Component::CORE), 'https://hyvor.mycompany.com');
+        $this->assertEquals(Component::BLOGS->getUrlOf(Component::TALK), 'https://talk.hyvor.mycompany.com');
     }
 
     public function testPlans(): void
     {
         $plans = [
-            [ComponentType::CORE, CorePlan::class],
-            [ComponentType::TALK, TalkPlan::class],
-            [ComponentType::BLOGS, BlogsPlan::class],
+            [Component::CORE, CorePlan::class],
+            [Component::TALK, TalkPlan::class],
+            [Component::BLOGS, BlogsPlan::class],
         ];
 
         foreach ($plans as $plan) {